Moved the base client and base client exception into an "Internals" folder so they dont clutter up the project root.
